fix: move step to utlis
